requirements.txt
streamlit
openai
langchain
python-dotenv
PyPDF2
langchain_communitys